Add get_file_string_length for raw data
While the protocol we are using is in raw mode while sending files
it is still a line based protocol.

As such, it cares much less about the file system size of the file and
much more so about the actual length of the raw string that is going to
be sent as bytes.

I'm not sure if there could be a data encoding issue, but I believe ASCII
is being used.

There should be no extra protocol nonsense embedded with the string.

Also add a few tests to show that I **think** this is case. But I could
certainly be incorrect