Start using deferred in the message processing
Use a message handler to handle the unique messages and allow line
receiver to just receive lines.

The lineReceiver method now sets up a deferred to handle the messages.

If it is working correctly, and there is an error handling the message,
then the _handleError (NAME WHICH WILL BE CHANGED) errBack will get called