Add tests for FileRequest, Mocks for Downloader, IO
The filerequest object has unit tests. The code has been changed to allow for
the injection of both a downloader and an IOHandler.

FileRequest.getFiles has been tested for functionality. It has not been tested
for file creation or file download.

Also, I have made comments on the download agent code. It seems that instance
variables might want to be changed. These could be parameters passed to
getFile. Lastly, work on how to get transfer status/progress out of the
protocol.